# specs/001-financius-web-companion/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Financius Web API
  version: 1.0.0
  description: |
    Versioned REST contract for Financius Web.
    All responses use the envelope: { data, error, meta }.
servers:
  - url: http://localhost:8000
    description: Local
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login with email/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /api/v1/auth/google:
    post:
      tags: [auth]
      summary: Login with Google OAuth code/token exchange
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: Logged in via Google
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
...
  /api/v1/receipts:
    get:
      tags: [receipts]
      summary: List receipts with search/filter/pagination
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
        - $ref: '#/components/parameters/Currency'
        - $ref: '#/components/parameters/CategoryId'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Receipt list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptListEnvelope'
  /api/v1/receipts/{receiptId}:
    get:
      tags: [receipts]
      summary: Get one receipt with line items
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      responses:
        '200':
          description: Receipt details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptEnvelope'
    patch:
      tags: [receipts]
      summary: Edit receipt fields
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptEnvelope'
  /api/v1/receipts/bulk:
    post:
      tags: [receipts]
      summary: Bulk mutate receipts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptBulkRequest'
      responses:
        '200':
          description: Bulk operation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResultEnvelope'
  /api/v1/receipts/import/amazon:
    post:
      tags: [receipts]
      summary: Import Amazon CSV receipts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResultEnvelope'
  /api/v1/categories:
    get:
      tags: [categories]
      summary: List category hierarchy
      responses:
        '200':
          description: Category tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryTreeEnvelope'
    post:
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
  /api/v1/categories/{categoryId}:
    patch:
      tags: [categories]
      summary: Update category
      parameters:
        - $ref: '#/components/parameters/CategoryIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
    delete:
      tags: [categories]
      summary: Delete category (with reassignment checks)
      parameters:
        - $ref: '#/components/parameters/CategoryIdPath'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
...
  /api/v1/shops/{shopId}/merge:
    post:
      tags: [shops]
      summary: Merge two shops
      parameters:
        - $ref: '#/components/parameters/ShopId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secondaryShopId]
              properties:
                secondaryShopId:
                  type: string
      responses:
        '200':
          description: Merge complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /api/v1/cards:
    get:
      tags: [cards]
      summary: List payment cards
      responses:
        '200':
          description: Card list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardListEnvelope'
    post:
      tags: [cards]
      summary: Create payment card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardEnvelope'
  /api/v1/budgets:
    get:
      tags: [budgets]
      summary: List budget entries for selected month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: '^\\d{4}-\\d{2}$'
      responses:
        '200':
          description: Budget entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetListEnvelope'
    post:
      tags: [budgets]
      summary: Create/update budget entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetUpsertRequest'
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    FromDate:
      name: from
      in: query
      schema: { type: string, format: date }
    ToDate:
      name: to
      in: query
      schema: { type: string, format: date }
    Currency:
      name: currency
      in: query
      schema:
        type: string
        enum: [USD, JPY, EUR, GBP, KRW, CNY, CAD, AUD]
    CategoryId:
      name: categoryId
      in: query
      schema: { type: string }
    Search:
      name: q
      in: query
      schema: { type: string }
...
    ReceiptId:
      name: receiptId
      in: path
      required: true
      schema: { type: string }
    CategoryIdPath:
      name: categoryId
      in: path
      required: true
      schema: { type: string }
    ShopId:
      name: shopId
      in: path
      required: true
      schema: { type: string }
  schemas:
    ApiEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          nullable: true
        error:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/MetaObject'
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    MetaObject:
      type: object
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
        pagination:
          type: object
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    GoogleLoginRequest:
      type: object
      required: [idToken]
      properties:
        idToken: { type: string }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    AuthPayload:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
    AuthEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthPayload'
...
    Receipt:
      type: object
      properties:
        id: { type: string }
        receiptDate: { type: string, format: date-time }
        shopName: { type: string }
        totalAmount: { type: number }
        currency: { type: string }
        categoryId: { type: string, nullable: true }
    ReceiptLineItem:
      type: object
      properties:
        name: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        lineTotal: { type: number }
    ReceiptDetail:
      allOf:
        - $ref: '#/components/schemas/Receipt'
        - type: object
          properties:
            lineItems:
              type: array
              items: { $ref: '#/components/schemas/ReceiptLineItem' }
    ReceiptEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ReceiptDetail' }
    ReceiptListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/Receipt' }
    ReceiptUpdateRequest:
      type: object
      properties:
        categoryId: { type: string, nullable: true }
        note: { type: string, nullable: true }
        paymentCardId: { type: string, nullable: true }
    ReceiptBulkRequest:
      type: object
      required: [operation, receiptIds]
      properties:
        operation:
          type: string
          enum: [recategorize, reassign_card, delete]
        receiptIds:
          type: array
          items: { type: string }
        categoryId: { type: string }
        paymentCardId: { type: string }
    BulkResultEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                affectedCount: { type: integer }
    Category:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        parentId: { type: string, nullable: true }
        displayOrder: { type: integer }
        isEngel: { type: boolean }
        needsWants: { type: string, enum: [needs, wants] }
        isHousing: { type: boolean }
        isFixedExpense: { type: boolean }
    CategoryCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        parentId: { type: string, nullable: true }
        displayOrder: { type: integer }
        isEngel: { type: boolean }
        needsWants: { type: string, enum: [needs, wants] }
        isHousing: { type: boolean }
        isFixedExpense: { type: boolean }
    CategoryUpdateRequest:
      type: object
      properties:
        name: { type: string }
        parentId: { type: string, nullable: true }
        displayOrder: { type: integer }
        isEngel: { type: boolean }
        needsWants: { type: string, enum: [needs, wants] }
        isHousing: { type: boolean }
        isFixedExpense: { type: boolean }
    CategoryEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Category' }
    CategoryTreeEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Category' }
...
    Card:
      type: object
      properties:
        id: { type: string }
        nickname: { type: string }
        cardType: { type: string }
        network: { type: string }
        colorHex: { type: string }
        isActive: { type: boolean }
    CardCreateRequest:
      type: object
      required: [nickname, cardType]
      properties:
        nickname: { type: string }
        cardType:
          type: string
          enum: [credit, debit, prepaid, digital_wallet]
        network: { type: string }
        colorHex: { type: string }
    CardEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Card' }
    CardListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Card' }
    Budget:
      type: object
      properties:
        id: { type: string }
        month: { type: string }
        categoryId: { type: string, nullable: true }
        mode:
          type: string
          enum: [manual, forecast, forecast_adjusted]
        amount: { type: number }
        adjustmentPct: { type: integer, nullable: true }
        rolloverEnabled: { type: boolean }
        rolloverBalance: { type: number }
    BudgetUpsertRequest:
      type: object
      required: [month, mode, amount]
      properties:
        month: { type: string, pattern: '^\\d{4}-\\d{2}$' }
        categoryId: { type: string, nullable: true }
        mode:
          type: string
          enum: [manual, forecast, forecast_adjusted]
        amount: { type: number }
        adjustmentPct: { type: integer, minimum: 80, maximum: 120 }
        rolloverEnabled: { type: boolean }
    BudgetEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Budget' }
    BudgetListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Budget' }
...